more organization and added route handlers for each model
