+++ server/item/models/Item.js
@@ -1,5 +1,5 @@
 const { Sequelize, DataTypes } = require("sequelize");
-const db = require("../config/db");
+const db = require("../../config/db");
 
 const Item = db.define(
   "item",
