--- server/participant/models/meeting.js
@@ -1,31 +1,0 @@
-const { Sequelize, DataTypes } = require("sequelize");
-const db = require("./config/db");
-
-const Meeting = db.define(
-  "meeting",
-  {
-    // Model attributes are defined here
-    id: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      primaryKey: true,
-    },
-    modId: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-    },
-    adminId: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-    },
-  },
-  {
-    freezeTableName: true,
-    timestamps: false,
-  }
-);
-
-// `sequelize.define` also returns the model
-console.log(Meeting === db.models.Meeting); // true
-
-module.exports = Meeting;
