+++ server/server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 // var fallback = require('express-history-api-fallback')
 const app = express();
+
 const path = require("path");
 const port = process.env.PORT || 3000;
 // import routes from "../client/routes";
@@ -10,7 +11,16 @@ const HTML_FILE = path.join(DIST_DIR, "index.html");
 
 const db = require("./config/db");
 
-const Participant = require("./models/participant");
+const Participant = require("./participant/models/participant");
+
+const participantController = require("./participant/participant-controller");
+
+const moderatorController = require("./moderator/moderator-controller");
+
+const meetingController = require("./meeting/meeting-controller");
+
+const itemController = require("./item/item-controller");
+const administratorController = require("./admin/admin-controller");
 
 try {
   db.authenticate().then(() => {
@@ -26,6 +36,12 @@ try {
 app.use(express.json());
 app.use(express.static(path.join(__dirname, "../public")));
 app.use(express.static(path.join(__dirname, "../dist")));
+
+app.use("/participants", participantController);
+app.use("/moderators", moderatorController);
+app.use("/meetings", meetingController);
+app.use("/items", itemController);
+app.use("/admins", administratorController);
 
 // app.use(express.static("helper"));
 // app.use("/", indexRoutes)
